refactor: extract tax advantaged account linking
